Adding TimeChart as a component
